Fix 80% warning

+++ 05-SceneManager/MoneyEffect.cpp
@@ -1,5 +1,5 @@
 #include "MoneyEffect.h"
-MoneyEffect::MoneyEffect(int pos_x, int pos_y, int moneyAni)
+MoneyEffect::MoneyEffect(float pos_x, float pos_y, int moneyAni)
 {
 	this->x = pos_x;
 	this->y = pos_y;
